backend/requirements.txt
fastapi
pypdf
uvicorn
python-multipart
langchain
langchain_core
langchain_openai
langchain_community
langchain_weaviate
weaviate-client